Clarify language in the README
When @outofambit and I walked through the repo earlier today, we thought this was a welcome page for the private org. 

I only just realized that this repo is publicly visible, which explains why info on how to get started as a member is fairly sparse.

Tried to make this more clear in the README so that new members don't go to this page and feel left in the dark. We'll also send them a welcome email upon adding them to the org to give them some first steps.